create dynamic to-json module
